Fix edge case behaviors for gnc
Now understands localhost and 127.0.0.1 are not remotes and will not use ssh.

Correctly runs services on localhost if mixed with services running on remotes.

Correctly runs when no services are specified, whether on local or remote.

Correctly runs when local services are given while no services are given for remote (i.e. gnc stop px4 @remotehost).